gh-634 Move TestDataFieldDefinition into TestDataGenerator
